requirements.txt
flask
python-dotenv
spotipy
requests

pip install -r requirements.txt

